based the container on Ubuntu
A lot of issues (wine version, basically) led to rethink
the container. It is now based on Ubuntu, which has simplified
the wine setup a little bit. (no more need to manually install
.net).

The created container comes with a pre-hacked STL plugin,
that be installed via the Sketchup Extension Manager.

Unfortunately, with wine 4.3, it does not seem to be possible
to symlink to system.reg and user.reg, because when accessing
to these files, the links are deleted and replaced by the plain
files.
The side effect is that the user has to always select a template,
and acknowledge the availabity of an update when starting the
container.

Signed-off-by: Thierry Bultel <thierry.bultel@iot.bzh>